requirements.txt
nextcord==2.6.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1